[ADD] Show Qty in Transaction UoM to Point of sales's orders analysis report
The Orders Analysis report showed in Point of Sales menu show the
product quantity in the default UoM. This information could be really
confusing when reviewing the report.

We change the Product Qty column name to Qty in Default UoM and add a
new column named Qty in Transaction UoM that show the quantity related
to the sale itself: the product unit of measure without any conversion.

This will will be easy to understand. Also add the Unit of Measure group
by to be able to filter the report using both in row/column Unit of
Measure group by

Also the product_qty field (Qty in Default UoM) is now a float field.
This way if the Transaction amount is smaller than the Default UoM will
represent the correct decimal number part.