[IMP] base_address_city: Improved way as append the city_id
Like the address_format could be changed by each country, this method
is required to append the field city_id in the views, to all country
address formats.

But if We need inherit the view to add a parameter in city or city_id,
WE need overwrite this method. Now, with this change, We only inherit
the view, and not all the method.